Hack for incorrect Const locations as described on -hackers. Tweaks.
Make fuzzing regression test with deeply nested set operation query not take so
long, without making it less useful. A 32-bit hash value is sufficient.

+++ src/include/nodes/plannodes.h
@@ -68,7 +68,7 @@ typedef struct PlannedStmt
 
 	int			nParamExec;		/* number of PARAM_EXEC Params used */
 
-	uint64		queryId;		/* query identifier carried from query tree */
+	uint32		queryId;		/* query identifier carried from query tree */
 } PlannedStmt;
 
 /* macro for fetching the Plan associated with a SubPlan node */
